Implementing all return possibilities, now to implement creation.
